Add layer_map argument
This argument is useful if you make additions to a .yaml model config file and want to change how pre-trained weights are loaded into it accordingly.

For example, if you add a new layer before the final detection layer of yolov9-c.yaml, you may want to remap the detection layer of an already-trained model from 38 to 39 by passing in --layer_map 38,39. This will allow all the weights of the pre-trained model to be utilized and leave the new layer 38 a fresh slate.

This commit implements utils.general.map_dicts. It takes in a list of strings of the format "old_layer_number,new_layer_number" and does the appropriate substitutions in the dictionary passed in.

This function is used in train_dual.py. When loading pretrained weights, the map_dicts function is called on the .pt file's weight dictionary to remap layer numbers.