Fix "Transfered x/x items from {weights}" output
With this change, correctly output the fraction of weights loaded from the pre-trained weights file. Also show the number of weights in the current model that will be freshly initialized.

Previously, this would output the number of weights loaded from the pre-trained model "/" the number of weights in the current model as the fraction of weights loaded. However, this will be incorrect if new layers have been added to the current model.

With this change, we show separately the fraction of weights loaded from the .pt file and the number of weights in the current model that will be newly initialized (no corresponding weights in the .pt file).